// ClientApp/components/interfaces/Resource.tsx
//RESOURCE TYPE
export interface Resource {
    name: string;           //name of resource
    homePage: string;       //homepage link
    color1: string;         //color scheme of link
    color2: string;         //""
    color3: string;         //""
    rssLink: string;        //list of rss links
    highlight: string;      //highlight color
    link1Name: string;      //link name
    link1Ref: string;       //link reference link
    link2Name: string;
    link2Ref: string;
    link3Name: string;
    link3Ref: string;
    link4Name: string;
    link4Ref: string;
    link5Name: string;
    link5Ref: string;
    link6Name: string;
    link6Ref: string;
    logo: string;
}